Move the stream selection tests next to the other EDF recording tests
