+++ tests/test_on_data/ecephys/test_recording_interfaces.py
@@ -347,6 +347,43 @@ class TestEDFRecordingInterface(RecordingExtractorInterfaceTestMixin):
         pass
 
 
+class TestEDFRecordingInterfaceStreamSelection:
+    """A file whose signals were not all sampled at the same rate carries one stream per rate."""
+
+    file_path = str(ECEPHY_DATA_PATH / "edf" / "heterogeneous_offsets" / "same_unit_offsets_multirate.edf")
+
+    def test_get_stream_names(self):
+        stream_names = EDFRecordingInterface.get_stream_names(file_path=self.file_path)
+
+        assert stream_names == ["stream ((100.0,) Hz)", "stream ((1.0,) Hz)"]
+
+    def test_stream_name_is_required_for_a_multi_stream_file(self):
+        with pytest.raises(ValueError, match="several streams"):
+            EDFRecordingInterface(file_path=self.file_path)
+
+    def test_stream_name_selects_the_channels_of_its_stream(self):
+        interface = EDFRecordingInterface(file_path=self.file_path, stream_name="stream ((1.0,) Hz)")
+
+        assert list(interface.channel_ids) == ["Resp oro-nasal", "EMG submental", "Temp rectal", "Event marker"]
+        assert interface.recording_extractor.get_sampling_frequency() == 1.0
+
+    def test_channels_to_skip_applies_within_the_selected_stream(self):
+        interface = EDFRecordingInterface(
+            file_path=self.file_path,
+            stream_name="stream ((100.0,) Hz)",
+            channels_to_skip=["EOG horizontal"],
+        )
+
+        assert list(interface.channel_ids) == ["EEG Fpz-Cz", "EEG Pz-Oz"]
+
+    def test_get_available_channel_ids_takes_a_stream_name(self):
+        channel_ids = EDFRecordingInterface.get_available_channel_ids(
+            file_path=self.file_path, stream_name="stream ((100.0,) Hz)"
+        )
+
+        assert channel_ids == ["EEG Fpz-Cz", "EEG Pz-Oz", "EOG horizontal"]
+
+
 class TestIntanRecordingInterfaceRHS(RecordingExtractorInterfaceTestMixin):
     data_interface_cls = IntanRecordingInterface
     interface_kwargs = dict(file_path=ECEPHY_DATA_PATH / "intan" / "intan_rhs_test_1.rhs")
